fix: move files to components folder
안쓰는 이미지 파일 삭제 및 사용하는 이미지 파일들을 assets 폴더로 이동
