+++ public/index.html
@@ -4,20 +4,9 @@
 <head>
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <meta name="description" content="판다마켓은 일상의 모든 물건을 거래할 수 있는 중고거래 사이트입니다." />
-    <meta property="og:url" content="https://panda-market.com/" />
-    <meta property="og:title" content="판다마켓">
-    <meta property="og:image" content="img/og/panda-market-preview-image.png" />
-    <meta property="og:description" content="일상의 모든 물건을 거래해 보세요" />
-    <meta name="twitter:title" content="판다마켓" />
-    <meta name="twitter:card" content="summary_large_image" />
-    <meta name="twitter:image" content="img/og/panda-market-preview-image.png" />
-    <meta name="twitter:description" content="일상의 모든 물건을 거래해 보세요" />
+    <link rel="icon" href="./favicon.ico" type="image/x-icon" />
     <title>판다마켓</title>
-    <link rel="stylesheet" href="/styles/reset.css" />
-    <link rel="stylesheet" href="/styles/fonts.css" />
-    <link rel="stylesheet" href="/styles/main.css" />
-    <link rel="icon" href="/img/logo/favicon.ico" type="image/x-icon" />
+
 </head>
 
 <body>
