+++ src/components/Header/Header.jsx
@@ -1,5 +1,5 @@
 import NavigationLink from './NavigationLink';
-import styles from '../styles/Header.module.css';
+import styles from '../../styles/Header.module.css';
 
 function Header() {
     return (
