--- src/components/ProductItem.jsx
@@ -1,10 +1,0 @@
-//상품에 대한 정보를 보여주는 컴포넌트
-function ProductItem() {
-    return (
-        <>
-            <h2>임시 정보창</h2>
-        </>
-    );
-}
-
-export default ProductItem;
