--- src/components/ProductList.jsx
@@ -1,12 +1,0 @@
-import SearchBar from './SearchBar';
-import ProductItem from './ProductItem';
-import Pagenation from './Pagination';
-//상품 목록을 보여주는 컴포넌트
-function ProductList() {
-    return (
-        <>
-            <ProductItem />
-        </>
-    );
-}
-export default ProductList;
