+++ src/styles/Items.module.css
@@ -1,48 +1,3 @@
-.items-page {
-    padding: 20px;
-}
-
-.sort-dropdown {
-    margin-bottom: 20px;
-}
-
-.product-list {
+.wrapper {
     display: flex;
-    flex-wrap: wrap;
-    gap: 20px;
-}
-
-.product-item {
-    border: 1px solid #ddd;
-    padding: 10px;
-    flex: 1 1 calc(25% - 20px); /* Desktop */
-}
-
-.product-item img {
-    width: 100%;
-    height: auto;
-}
-
-.pagination {
-    display: flex;
-    justify-content: center;
-    margin-top: 20px;
-}
-
-.pagination button {
-    margin: 0 10px;
-}
-
-/* Tablet */
-@media (max-width: 1023px) {
-    .product-item {
-        flex: 1 1 calc(33.33% - 20px);
-    }
-}
-
-/* Mobile */
-@media (max-width: 767px) {
-    .product-item {
-        flex: 1 1 calc(50% - 20px);
-    }
 }
