+++ src/styles/NavigationLink.module.css
@@ -1,7 +1,7 @@
 .home {
     width: 153px;
     height: 51px;
-    background-image: url('/img/logo/panda-market-logo.png');
+    background-image: url('../assets/img/panda-market-logo.png');
     background-size: contain;
     background-repeat: no-repeat;
     text-indent: -9999px;
@@ -51,7 +51,7 @@ a.active {
     .home {
         width: 81px;
         height: 40px;
-        background-image: url('/img/logo/panda-market-logo-mobile.png');
+        background-image: url('../assets/img/panda-market-logo-mobile.png');
     }
 
     .nav {
